Restructure video component to allow dynamic selection of VLC or OMXPlayer based renderer on raspberry pi
