+++ es-app/src/views/gamelist/VideoGameListView.h
@@ -5,7 +5,6 @@
 #include "components/RatingComponent.h"
 #include "components/DateTimeComponent.h"
 #include "components/VideoComponent.h"
-#include "components/VideoPlayerComponent.h"
 
 class VideoGameListView : public BasicGameListView
 {
@@ -29,8 +28,7 @@ private:
 	void initMDValues();
 
 	ImageComponent mMarquee;
-	//VideoComponent mVideo;
-	VideoPlayerComponent mVideo;
+	VideoComponent* mVideo;
 	ImageComponent mImage;
 
 	TextComponent mLblRating, mLblReleaseDate, mLblDeveloper, mLblPublisher, mLblGenre, mLblPlayers, mLblLastPlayed, mLblPlayCount;
