+++ es-core/src/components/VideoComponent.h
@@ -8,17 +8,9 @@
 #include "ImageComponent.h"
 #include <string>
 #include <memory>
-#include "resources/TextureResource.h"
-#include <vlc/vlc.h>
 #include <SDL.h>
 #include <SDL_mutex.h>
 #include <boost/filesystem.hpp>
-
-struct VideoContext {
-	SDL_Surface*		surface;
-	SDL_mutex*			mutex;
-	bool				valid;
-};
 
 class VideoComponent : public GuiComponent
 {
@@ -32,8 +24,6 @@ class VideoComponent : public GuiComponent
 	};
 
 public:
-	static void setupVLC();
-
 	VideoComponent(Window* window);
 	virtual ~VideoComponent();
 
@@ -55,8 +45,6 @@ public:
 	void onSizeChanged() override;
 	void setOpacity(unsigned char opacity) override;
 
-	void render(const Eigen::Affine3f& parentTrans) override;
-
 	virtual void applyTheme(const std::shared_ptr<ThemeData>& theme, const std::string& view, const std::string& element, unsigned int properties) override;
 
 	virtual std::vector<HelpPrompt> getHelpPrompts() override;
@@ -66,35 +54,29 @@ public:
 
 	virtual void update(int deltaTime);
 
-private:
+	void render(const Eigen::Affine3f& parentTrans) override;
+
+protected:
 	// Start the video Immediately
-	void startVideo();
+	virtual void startVideo() = 0;
+	// Stop the video
+	virtual void stopVideo() { };
+	// Handle looping the video. Must be called periodically
+	virtual void handleLooping();
+
 	// Start the video after any configured delay
 	void startVideoWithDelay();
-	// Stop the video
-	void stopVideo();
-
-	void setupContext();
-	void freeContext();
 
 	// Handle any delay to the start of playing the video clip. Must be called periodically
 	void handleStartDelay();
 
-	// Handle looping the video. Must be called periodically
-	void handleLooping();
-
 	// Manage the playing state of the component
 	void manageState();
 
-private:
-	static libvlc_instance_t*		mVLC;
-	libvlc_media_t*					mMedia;
-	libvlc_media_player_t*			mMediaPlayer;
-	VideoContext					mContext;
+protected:
 	unsigned						mVideoWidth;
 	unsigned						mVideoHeight;
 	Eigen::Vector2f 				mOrigin;
-	std::shared_ptr<TextureResource> mTexture;
 	float							mFadeIn;
 	std::string						mStaticImagePath;
 	ImageComponent					mStaticImage;
